+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [Codelist] Extend `plu:docType` to describe values from https://www.xrepository.de/details/urn:xoev-de:xplanverfahren:codeliste:verfahrensunterlagentyp
 * [Website] Fix links to DCAT-AP.PLU documentation
+* [Deployment] Set draft version for development
 
 ## 2024-02-28 - 0.1.4